InboxOracle/requirements.txt
google-auth>=2.22.0
google-api-python-client>=2.95.0
google-auth-oauthlib>=1.0.0
pytest>=7.4.0
python-dateutil>=2.8.2
SQLAlchemy>=2.0.19
loguru>=0.7.0